Move ClusterAutoKResult to types
